Add OOGABOOGA_HEADLESS option
